refactor(rm-to-trash): コードを整理
src/main.rsを整理して、
I/O処理やシリアライズ処理や、
書き換え処理を分離。

またこれまで予想外だった普通は起きない状態でも黙って正常終了していましたが、
panicにしてユーザーに知らせて修正を促すようにしました。
それに伴い`main`関数はExitCodeを返すのではなく、
何も返さない関数になっています。

READMEとコードコメントで説明が重複しているような部分を消して、
コードとテストを見たほうが正確に分かりやすいようなコメントも抽象的にしました。

ソースコードモジュールが複数に増えたので、
`codegen-units`を1にしたことはあまり無料ではなくなったと思ったので、
デフォルトに任せます。
どちらにしてもこの程度のプログラムでは大したビルド時間にはならないと思いますが。
