+++ plugins/rm-to-trash/Cargo.toml
@@ -10,6 +10,5 @@ serde_json = "1.0.149"
 winnow = "1.0.2"
 
 [profile.release]
-codegen-units = 1
 lto = true
 strip = true
