+++ plugins/rm-to-trash/src/mod.rs
@@ -1,0 +1,2 @@
+pub mod crate::rewrite;
+pub mod crate::serialize;
